🚚 [projects] Move function `continueupdate` from `services` to `update` module
