+++ src/cutty/services/update.py
@@ -44,20 +44,6 @@ def update(
     updateproject(projectdir, createproject)
 
 
-def continueupdate(projectdir: Path) -> None:
-    """Continue an update after conflict resolution."""
-    project = Repository.open(projectdir)
-
-    if commit := project.cherrypickhead:
-        project.commit(
-            message=commit.message,
-            author=commit.author,
-            committer=project.default_signature,
-        )
-
-    project.heads[LATEST_BRANCH] = project.heads[UPDATE_BRANCH]
-
-
 def skipupdate(projectdir: Path) -> None:
     """Skip an update with conflicts."""
     project = Repository.open(projectdir)
